move zkapp proving logic to separate file
